refactor(arch): reorganise utils/ into core/ and rendering/ subpackages
context.py moves to core/ (it is a core domain concern — context
detection drives workspace/org resolution). rich_tables.py and
table_renderer.py move to a new rendering/ subpackage (a coherent
rendering layer, not a utility grab-bag).

utils/ now contains only leaf utilities with no internal terrapyne
dependencies: change_directory, browser helpers, and logging
formatters.

Three new importlinter contracts enforce the boundaries:
- Utils Has No Upward Imports: utils/ cannot import cli/api/core/models/rendering
- Rendering Imports Only Models And Utils: rendering/ cannot import api/core
- CLI Isolation updated to include rendering/ in the forbidden set

All 609 tests pass; ruff and mypy clean.
