+++ src/terrapyne/rendering/__init__.py
@@ -1,0 +1,1 @@
+"""Rendering subpackage — Rich table formatters for TFC data."""
