+++ src/terrapyne/rendering/table_renderer.py
